+++ src/main/java/survival2d/service/domain/LobbyTeam.java
@@ -9,10 +9,10 @@ import lombok.Getter;
 public class LobbyTeam implements Team {
 
   @Getter
-  private final long id;
+  private final int id;
   private final Set<String> playerUsernames = new HashSet<>(GameConstant.TEAM_PLAYER);
 
-  public LobbyTeam(long teamId) {
+  public LobbyTeam(int teamId) {
     id = teamId;
   }
